refactor(code): refactor the code for the linting and typos
