+++ pkg/generic/pod-autoscaler/types/types.go
@@ -21,6 +21,6 @@ type ExperimentDetails struct {
 	ChaosPodName     string
 	RunID            string
 	AuxiliaryAppInfo string
-	Timeout             int
-	Delay               int
+	Timeout          int
+	Delay            int
 }
