+++ pkg/math/math.go
@@ -16,7 +16,7 @@ func Minimum(a int, b int) int {
 	return a
 }
 
-//Rule of three for calculating an integer given another integer representing a percentage
+//Adjustment contains rule of three for calculating an integer given another integer representing a percentage
 func Adjustment(a int, b int) int {
 
 	b = a * b / 100
